requirements.txt
aiohttp>=3.8.0
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pathlib>=1.0.1
click>=8.0.0
rich>=10.0.0
aiofiles>=0.8.0
pyyaml>=6.0.0
jsonschema>=4.0.0
backoff>=2.2.0
tenacity>=8.0.0
